feat: finalizão da tela

+++ src/pages/descontoCliente/descontoCliente.ts
@@ -1,48 +1,28 @@
 import { reactive } from "vue";
+import $ from "jquery";
 import xGridV2, { ixGridCreate } from "@/plugins/xGridV2";
 import xModal, { iModalCreate } from '@/plugins/xModal/xModal';
 import Swal from "sweetalert2";
-import { msgConfirm } from "@/ts/message";
-
 import serviceDescontoCliente from "./services/descontoCliente.service"
-
-import { iCliente, iMarca, iMarcaAdicionada, iParamGetMarcas, iParamGetMarcasAdicionadas } from "./interfaces";
+import { iCliente, iMarca, iMarcaAdicionadaForm, iParamGetMarcas, iParamGetMarcasAdicionadas } from "./interfaces";
 import utils from "@/ts/utils";
-
-interface _iMarcaxGridCreate extends ixGridCreate {
-    dataSource: (obj?: object) => iMarca;
-}
-
-interface _iMarcaAdicionadaxGridCreate extends ixGridCreate {
-    dataSource: (obj?: object) => iMarcaAdicionada;
-}
+import { msgConfirm } from "@/ts/message";
 
 export const state = reactive({
     gridCliente: <ixGridCreate>{},
-    gridMarca: <_iMarcaxGridCreate>{},
-    gridMarcaAdicionada: <_iMarcaAdicionadaxGridCreate>{},
+    gridMarca: <ixGridCreate>{},
+    gridMarcaAdicionada: <ixGridCreate>{},
+
     edtMarcaSearch: <HTMLInputElement>{},
 
     modalCliente: <iModalCreate>{},
     modalAdicionarMarca: <iModalCreate>{},
 
-    dbCliente: <iCliente>{},
     dbClienteSelecionado: <iCliente>{},
-
     dbMarca: <iMarca>{},
 
-    dbMarcaAdicionada: <iMarcaAdicionada>{},
-
     loading: false,
-    modalClienteOpened: false,
-    modalAdicionarMarcaOpened: false,
-    pnSearch: false,
-
-    configVMoney: {
-        thousands: '.',
-        decimal: ',',
-        precision: 2,
-    }
+    pnSearch: true,
 })
 
 export const actions = {
@@ -64,45 +44,25 @@ export const actions = {
             },
         }),
 
-        state.gridMarcaAdicionada = new xGridV2.create({
-            el: "#pnMarcasAdicionadas",
-            height: 400,
-            columns: {
-                "Marca": { dataField: "DESCRICAO", width: "45%" },
-                "Desconto %": { dataField: "DESCONTO", render: utils.formatValor, width: "16%", center: true },
-                "Data Inicial": { dataField: "DATA_INICIAL", render: utils.dataBrasil, center: true },
-                "Data Final": { dataField: "DATA_FINAL", render: utils.dataBrasil, center: true },
-            },
-            query: {
-                async execute(rs) {
-                    let data = await actions.getMarcasAdicionadas({
-                        offset: rs.offset,
-                        param: rs.param
-                    })
-                    state.gridMarcaAdicionada.querySourceAdd(data)
-                }
-            },
-            sideBySide: {
-                el: "#marcasAdicionadasCampos",
-                frame: {
-                    el: "#btnMarcasAdicionadas",
-                    buttons: {
-                        Cancel: {
-                            html: "Cancelar",
-                            state: "select",
-                            click: actions.btnCancelAdicionarMarca,
-                            preLoad: "Cancelando o cliente",
-                        },
-                        Ok: {
-                            html: "OK",
-                            state: "select",
-                            click: actions.btnSelectAdicionarMarca,
-                            preLoad: "Selecionando o cliente",
-                        },
+            state.gridMarcaAdicionada = new xGridV2.create({
+                el: "#pnMarcasAdicionadas",
+                height: 400,
+                columns: {
+                    "Marca": { dataField: "DESCRICAO", width: "45%" },
+                    "Desconto %": { dataField: "DESCONTO", render: utils.formatValor, width: "16%", center: true },
+                    "Data Inicial": { dataField: "DATA_INICIAL", render: utils.dataBrasil, center: true },
+                    "Data Final": { dataField: "DATA_FINAL", render: utils.dataBrasil, center: true },
+                },
+                query: {
+                    async execute(rs) {
+                        let data = await actions.getMarcasAdicionadas({
+                            offset: rs.offset,
+                            param: rs.param
+                        })
+                        state.gridMarcaAdicionada.querySourceAdd(data)
                     }
-                }
-            }
-        })
+                },
+            })
     },
 
     criarModais() {
@@ -114,11 +74,20 @@ export const actions = {
         }),
 
             state.modalAdicionarMarca = new xModal.create({
-                height: 215,
+                height: 221,
                 width: 340,
                 el: '#mdAdicionarMarca',
                 theme: "xModal-blue",
             })
+    },
+
+    init() {
+        $(".ss").attr("autocomplete", "off");
+
+        state.edtMarcaSearch = <any>document.getElementById("edtMarcaSearch");
+
+        actions.grids();
+        actions.criarModais();
     },
 
     onClickClienteModal() {
@@ -130,7 +99,23 @@ export const actions = {
     },
 
     onClickAdicionarMarcaModal() {
+        const marca = state.gridMarca.dataSource()
+
+        if (!marca) {
+            Swal.fire({
+                icon: "warning",
+                title: "Selecione uma marca",
+            });
+            return;
+        }
+
+        state.dbMarca = marca
+
         state.modalAdicionarMarca.open()
+    },
+
+    onClickCloseAdicionarMarcaModal() {
+        state.modalAdicionarMarca.close()
     },
 
     searchMarcas() {
@@ -151,6 +136,8 @@ export const actions = {
         state.gridMarcaAdicionada.queryOpen({ ID_CLIENTE: state.dbClienteSelecionado.ID_CLIENTE }, () => {
             state.gridMarcaAdicionada.focus();
         });
+
+        state.pnSearch = false;
 
         state.modalCliente.close()
     },
@@ -191,6 +178,47 @@ export const actions = {
         }
     },
 
+    async adicionarMarca(marcaAdicionada: iMarcaAdicionadaForm) {
+        try {
+            const marca = state.gridMarca.dataSource()
+
+            let idMarca = marca.ID_MARCA
+            let idCliente = state.dbClienteSelecionado.ID_CLIENTE
+            let descricao = marca.DESCRICAO
+            let desconto = parseFloat(marcaAdicionada.DESCONTO.replace(',', '.')) * 10  
+
+            if (desconto > 100) {
+                desconto = 100 
+            }
+            
+            let dadosMarcaAdicionada = {
+                ID_MARCA: idMarca,
+                ID_CLIENTE: idCliente,
+                DESCONTO: desconto,
+                DESCRICAO: descricao,
+                DATA_INICIAL: marcaAdicionada.DATA_INICIAL,
+                DATA_FINAL: marcaAdicionada.DATA_FINAL
+            }
+
+            state.loading = true
+            const data = await serviceDescontoCliente.adicionarMarca(dadosMarcaAdicionada)
+            state.loading = false
+
+            state.modalAdicionarMarca.close()
+            state.gridMarcaAdicionada.insertLine(dadosMarcaAdicionada)
+            state.gridMarca.deleteLine()
+
+            return data
+
+        } catch (error) {
+            state.loading = false
+            Swal.fire({
+                icon: "error",
+                text: "Erro ao adicionar a marca!"
+            })
+        }
+    },
+
     async removerMarca() {
         try {
 
@@ -207,14 +235,17 @@ export const actions = {
             let idMarca = marcaAdicionada.ID_MARCA
             let idCliente = state.dbClienteSelecionado.ID_CLIENTE
 
-            state.loading = true
-            const data = await serviceDescontoCliente.removerMarca(idMarca, idCliente)
-            state.loading = false
+            if (await msgConfirm("Confirmação", "Confirma a exclusão desta marca?")) {
+                state.loading = true
+                const data = await serviceDescontoCliente.removerMarca(idMarca, idCliente)
+                state.loading = false
 
-            state.gridMarcaAdicionada.deleteLine();
-            state.gridMarca.insertLine(marcaAdicionada);
+                state.gridMarcaAdicionada.deleteLine();
+                state.gridMarca.insertLine(marcaAdicionada);
 
-            return data
+                return data
+            }
+
         } catch (error) {
             state.loading = false
             Swal.fire({
